lisää createpagen refaktorointia
