+++ frontend/src/pages/createpage.jsx
@@ -4,6 +4,7 @@ import LoginPage from "./loginpage";
 import "../index.css";
 import newaccountcheck from "../utils/newaccountcheck";
 import CreateForm from "../../components/CreateForm";
+import createaccount from "../utils/createaccount";
 
 const NewAccountPage = ({ onAccountCreated }) => {
   const [username, setUsername] = useState("");
@@ -30,57 +31,23 @@ const NewAccountPage = ({ onAccountCreated }) => {
     if (typeof response === "string") {
       setError(response);
     } else if (response === true) {
-      createAccount();
-    }
-  };
-
-  const createAccount = () => {
-    /*
-    Send request to server to check if email is already in use
-    */
-
-    axios
-      .get(`${API_URL}/api/listobjects/users/?email=${email}`)
-      .then((response) => {
-        console.log(response);
-        console.log(response.data);
-        const existingUsers = response.data;
-        if (existingUsers.some((user) => user.email === email)) {
-          setError("Sähköposti on jo käytössä.");
-        } else {
-          const userObject = {
-            username,
-            password,
-            email,
-            telegram,
-            role: 5,
-            organization: null,
-            keys: null,
-          };
-          axios
-            .post(`${API_URL}/api/users/register`, userObject)
-            .then((response) => {
-              console.log(response);
-              console.log("Account created successfully!");
-              setUserCreated(true);
-              onAccountCreated && onAccountCreated();
-
-              // Set timeout to hide success message after 5 seconds
-              setTimeout(() => {
-                setUserCreated(false);
-              }, 5000);
-            })
-            .catch((error) => {
-              console.error("Error creating account:", error);
-              setError("Virhe luotaessa käyttäjää.");
-            });
-          setShowLoginPage(true);
-        }
-      })
-      .catch((error) => {
-        console.error("Error checking email:", error);
-        setError("Virhe tarkistettaessa sähköpostia.");
+      const resp = await createaccount({
+        API_URL,
+        email,
+        username,
+        password,
+        telegram,
+        setUserCreated,
+        setShowLoginPage,
+        onAccountCreated,
       });
+      console.log(resp);
+      if (typeof resp === "string") {
+        setError(resp);
+      }
+    } else {
+      setError("Virhe käyttäjän luonnissa.");
+    }
   };
 
   const handleBackToLogin = () => {
